Refactor according to the comments:
Notes on Mauricio Arellano implementation.
 
·         Can we separate the entities ( Pojos ) and the service classes
, all are inside the same package now.
·         Can we have some interface defined for Auction, Right now we
have none.
·         In general, exceptions are used only to report error
scenarios, the usage of MaxBidReachedException to report Maximum Bid
getting reached by a bidder is not appropriate.
A bidder reaching their maximum limit is a natural outcome of the
bidding processes and should not used as means to communicate a natural
event.
There are 2 usage of this and both of them should be corrected.
 
Case 1 : Auction.java
 
            if(currentBid > bidder.getOffer().getMaxBid()){
                        throw new MaxBidReachedException();
      }
 
Case 2: Bidder.java
 
            if(bid == nexBid){
                        throw new MaxBidReachedException();
            }
 
Can we have these corrected first, Thanks.